backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
python-dotenv
pydantic
chromadb
psycopg2-binary
sqlalchemy
mcp
httpx
PyJWT
python-multipart
# Phase 2: Communication Integrations
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
slack-bolt
slack-sdk
apscheduler
